decouple useGetVaults
