+++ src/clients/api/queries/useGetVaults/constants.ts
@@ -1,0 +1,5 @@
+import { getToken, getContractAddress } from 'utilities';
+
+export const XVS_TOKEN_ID = 'xvs';
+export const XVS_TOKEN_ADDRESS = getToken(XVS_TOKEN_ID).address;
+export const XVS_VAULT_ADDRESS = getContractAddress('xvsVault');
