+++ src/clients/api/queries/useGetVaults/useGetXvsVaultPoolBalances.ts
@@ -1,0 +1,42 @@
+import { useQueries, UseQueryOptions, UseQueryResult } from 'react-query';
+
+import { useWeb3 } from 'clients/web3';
+import { getTokenContractByAddress } from 'clients/contracts';
+import FunctionKey from 'constants/functionKey';
+import { Bep20 } from 'types/contracts';
+import { XVS_TOKEN_ID, XVS_VAULT_ADDRESS } from './constants';
+import getBalanceOf, { GetBalanceOfOutput } from '../getBalanceOf';
+
+export interface UseGetXvsVaultPoolBalancesInput {
+  stakedTokenAddresses: (string | undefined)[];
+}
+
+export type UseGetXvsVaultPoolBalancesOutput = UseQueryResult<GetBalanceOfOutput>[];
+
+const useGetXvsVaultPoolBalances = ({
+  stakedTokenAddresses,
+}: UseGetXvsVaultPoolBalancesInput): UseGetXvsVaultPoolBalancesOutput => {
+  const web3 = useWeb3();
+
+  // Query total amount of tokens staked for each pool
+  const queries: UseQueryOptions<GetBalanceOfOutput>[] = stakedTokenAddresses.map(
+    stakedTokenAddress => {
+      const tokenContract =
+        stakedTokenAddress && getTokenContractByAddress(stakedTokenAddress, web3);
+
+      return {
+        queryFn: () =>
+          getBalanceOf({
+            tokenContract: tokenContract || ({} as Bep20),
+            accountAddress: XVS_VAULT_ADDRESS,
+          }),
+        queryKey: [FunctionKey.GET_BALANCE_OF, XVS_TOKEN_ID, XVS_VAULT_ADDRESS],
+        enabled: !!tokenContract,
+      };
+    },
+  );
+
+  return useQueries(queries);
+};
+
+export default useGetXvsVaultPoolBalances;
